+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.16.9
 
 * Colorful output and a new cli argument to disable color (--no-color) #191
+* Support multi line environment variables from script output #196
 
 ### v0.16.8 (2019-02-09)